Separated components into separate files & added comments
