+++ Carbon/navigation/screens/Questionnaire/QuestionMileage.js
@@ -1,0 +1,78 @@
+import React, {useState} from 'react';
+import {View, Text,Button,TextInput } from 'react-native';
+import { Colors } from '../../../colors/Colors';
+
+/*
+Mileage Screen
+
+TODO: Improve UI
+TODO: Improve transferring of data between pages
+*/
+
+export default function MileageScreen({navigation,route}) {
+    //Values from previous pages
+    const dietScore = route.params?.dietScore;
+    const homePowerScore = route.params?.homePowerScore;
+    const annualPower = route.params?.annualPower;
+    const transportScore= route.params?.transportScore;
+
+    //Value to calculate & transfer at the "finished" screen
+    const [miles,setMiles] = useState(0);
+
+    return (
+    <>
+    <View
+    style={{
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+    width: "100%"
+    }}
+    >
+        <Text>
+        (Optional) What is the mileage on your vehicle?
+        </Text>
+        <View>
+        <Text> Miles </Text>
+        <TextInput
+        placeholder="Ex: 208309"
+        style={{
+        backgroundColor:Colors.secondary.NYANZA,
+        }}
+        keyboardType="decimal-pad"
+        onChangeText={text=>setMiles(text)}
+        />
+        </View>
+        </View>
+            <View style={{
+                justifyContent:'center',
+                flexDirection:"row",
+            }}>
+            <View style={{width:'50%'}}>
+            <Button
+            title="Previous Question"
+            color={Colors.primary.MINT}
+            onPress={() =>
+                navigation.goBack()
+            }
+            />
+            </View>
+            <View style={{width:'50%'}}>
+            <Button
+            title="Next Question"
+            color={Colors.primary.MINT}
+            onPress={() =>
+                navigation.navigate('finished',{
+                    transportScore:transportScore,
+                    dietScore:dietScore,
+                    homePowerScore:homePowerScore,
+                    annualPower:annualPower,
+                    miles:miles,
+                    })
+            }
+            />
+            </View>
+            </View>
+            </>
+        )
+}
